// src/app/models/change-pin-om.model.ts
export interface ChangePinOm {
	msisdn?: string;
	em?: string;
	pin?: string;
	new_pin?: string;
	confirm_pin?: string;
	code_otp?: string;
	hmac?: string;
	uuid?: string;
	os?: string;
	channel?: string;
	firebase_id?: string;
	app_version?: string;
	conf_version?: string;
	service_version?: string;
	type?: string;
	is_primo?: boolean;
}